+++ packages/adapter-kysely/src/utils.ts
@@ -2,7 +2,7 @@ import type { KeySchema, SessionSchema } from "lucia-auth";
 import type { Selectable } from "kysely";
 import type { KyselyKey, KyselySession } from "./types.js";
 
-export const convertSessionData = (
+export const transformSessionData = (
 	session: Selectable<KyselySession>
 ): SessionSchema => {
 	return {
@@ -15,7 +15,7 @@ export const convertSessionData = (
 
 export type Dialect = "pg" | "mysql2" | "better-sqlite3";
 
-export const convertKeyData = (key: Selectable<KyselyKey>): KeySchema => {
+export const transformKeyData = (key: Selectable<KyselyKey>): KeySchema => {
 	return {
 		id: key.id,
 		user_id: key.user_id,
@@ -25,7 +25,7 @@ export const convertKeyData = (key: Selectable<KyselyKey>): KeySchema => {
 	};
 };
 
-export const convertKeySchemaToKyselyValues = (
+export const transformKeySchemaToKyselyExpectedValue = (
 	key: KeySchema,
 	dialect: Dialect
 ): Selectable<KyselyKey> => {