+++ packages/adapter-prisma/src/utils.ts
@@ -1,6 +1,6 @@
 import type { KeySchema, SessionSchema } from "lucia-auth";
 
-export const convertSessionData = (
+export const transformSessionData = (
 	sessionData: SessionSchema
 ): SessionSchema => {
 	const { active_expires, idle_expires: idleExpires, ...data } = sessionData;
@@ -11,7 +11,7 @@ export const convertSessionData = (
 	};
 };
 
-export const convertKeyData = (keyData: KeySchema): KeySchema => {
+export const transformKeyData = (keyData: KeySchema): KeySchema => {
 	const { expires, ...data } = keyData;
 	return {
 		expires: expires === null ? null : Number(expires),