+++ packages/integration-oauth/src/providers/index.ts
@@ -1,0 +1,8 @@
+export { google, type GoogleUser } from "./google.js";
+export { github, type GithubUser } from "./github.js";
+export { discord, type DiscordUser } from "./discord.js";
+export { patreon, type PatreonUser } from "./patreon.js";
+export { twitch, type TwitchUser } from "./twitch.js";
+export { reddit, type RedditUser } from "./reddit.js";
+export { linkedin, type LinkedinUser } from "./linkedin.js";
+export { auth0, type Auth0User } from "./auth0.js";