+++ packages/lucia-auth/src/auth/index.ts
@@ -1,45 +1,75 @@
-import type {
-	Env,
-	Key,
-	MinimalRequest,
-	Session,
-	SessionSchema,
-	User
-} from "../types.js";
 import { Cookie, CookieOption, createSessionCookie } from "./cookie.js";
-import {
-	Adapter,
-	SessionAdapter,
-	UserAdapter,
-	UserData,
-	UserSchema
-} from "../types.js";
 import { logError } from "../utils/log.js";
 import {
 	generateHashWithScrypt,
 	generateRandomString,
 	validateScryptHash
 } from "../utils/crypto.js";
-import { LuciaError, LuciaErrorConstructor } from "./error.js";
+import { LuciaError } from "./error.js";
 import { parseCookie } from "../utils/cookie.js";
-import { transformDatabaseSessionData } from "./session.js";
-import { transformDatabaseKeyData, getOneTimeKeyExpiration } from "./key.js";
-
+import { validateDatabaseSessionData } from "./session.js";
+import { transformDatabaseKey, getOneTimeKeyExpiration } from "./key.js";
 export { SESSION_COOKIE_NAME } from "./cookie.js";
 
-export const lucia = <C extends Configurations>(configs: C) => {
-	return new Auth(configs);
+import type { UserSchema, SessionSchema, KeySchema } from "./schema.type.js";
+import type { Adapter, UserAdapter, SessionAdapter } from "./adapter.type.js";
+import type { LuciaErrorConstructor } from "./error.js";
+import { isWithinExpiration } from "../utils/date.js";
+
+export type Session = Readonly<{
+	sessionId: string;
+	userId: string;
+	activePeriodExpires: Date;
+	idlePeriodExpires: Date;
+	state: "idle" | "active";
+	isFresh: boolean;
+}>;
+
+export type Key = SingleUseKey | PersistentKey;
+
+export type SingleUseKey = Readonly<{
+	type: "single_use";
+	userId: string;
+	providerId: string;
+	providerUserId: string;
+	isPasswordDefined: boolean;
+	expires: Date;
+	isExpired: boolean;
+}>;
+
+export type PersistentKey = Readonly<{
+	type: "persistent";
+	userId: string;
+	providerId: string;
+	providerUserId: string;
+	isPasswordDefined: boolean;
+	isPrimary: boolean;
+}>;
+
+export type Env = "DEV" | "PROD";
+export type User = ReturnType<Lucia.Auth["transformUserData"]>;
+export type UserData = { id: string } & Required<Lucia.UserAttributes>;
+export type MinimalRequest = {
+	headers: {
+		get: (name: string) => null | string;
+	};
+	url: string;
+	method: string;
 };
 
-const validateConfigurations = (configs: Configurations) => {
-	const isAdapterProvided = configs.adapter;
+export const lucia = <C extends Configuration>(config: C) => {
+	return new Auth(config);
+};
+
+const validateConfiguration = (config: Configuration) => {
+	const isAdapterProvided = config.adapter;
 	if (!isAdapterProvided) {
 		logError('Adapter is not defined in configuration ("config.adapter")');
 		process.exit(1);
 	}
 };
 
-export class Auth<C extends Configurations = any> {
+export class Auth<C extends Configuration = any> {
 	private adapter: Adapter;
 	private generateUserId: () => MaybePromise<string>;
 	private sessionCookie: CookieOption[];
@@ -60,36 +90,41 @@ export class Auth<C extends Configurations = any> {
 		: { userId: string };
 	private csrfProtection: boolean;
 
-	constructor(configs: C) {
-		validateConfigurations(configs);
+	constructor(config: C) {
+		validateConfiguration(config);
 		const defaultSessionCookieOption: CookieOption = {
 			sameSite: "lax",
 			path: "/"
 		};
-		if ("user" in configs.adapter) {
-			if ("getSessionAndUserBySessionId" in configs.adapter.user) {
-				delete configs.adapter.user.getSessionAndUserBySessionId;
+		if ("user" in config.adapter) {
+			let userAdapter = config.adapter.user(LuciaError);
+			let sessionAdapter = config.adapter.session(LuciaError);
+			if ("getSessionAndUserBySessionId" in userAdapter) {
+				const { getSessionAndUserBySessionId: _, ...extractedUserAdapter } =
+					userAdapter;
+				userAdapter = extractedUserAdapter;
 			}
-			if ("getSessionAndUserBySessionId" in configs.adapter.session) {
-				delete configs.adapter.session.getSessionAndUserBySessionId;
+			if ("getSessionAndUserBySessionId" in sessionAdapter) {
+				const { getSessionAndUserBySessionId: _, ...extractedSessionAdapter } =
+					sessionAdapter;
+				sessionAdapter = extractedSessionAdapter;
 			}
+			this.adapter = {
+				...userAdapter,
+				...sessionAdapter
+			};
+		} else {
+			this.adapter = config.adapter(LuciaError);
 		}
-		this.adapter =
-			"user" in configs.adapter
-				? {
-						...configs.adapter.user(LuciaError),
-						...configs.adapter.session(LuciaError)
-				  }
-				: configs.adapter(LuciaError);
 		this.generateUserId =
-			configs.generateCustomUserId ?? (() => generateRandomString(15));
-		this.ENV = configs.env;
-		this.csrfProtection = configs.csrfProtection ?? true;
+			config.generateCustomUserId ?? (() => generateRandomString(15));
+		this.ENV = config.env;
+		this.csrfProtection = config.csrfProtection ?? true;
 		this.sessionTimeout = {
-			activePeriod: configs.sessionTimeout?.activePeriod ?? 1000 * 60 * 60 * 24,
-			idlePeriod: configs.sessionTimeout?.idlePeriod ?? 1000 * 60 * 60 * 24 * 14
+			activePeriod: config.sessionTimeout?.activePeriod ?? 1000 * 60 * 60 * 24,
+			idlePeriod: config.sessionTimeout?.idlePeriod ?? 1000 * 60 * 60 * 24 * 14
 		};
-		this.autoDatabaseCleanup = configs.autoDatabaseCleanup ?? true;
+		this.autoDatabaseCleanup = config.autoDatabaseCleanup ?? true;
 		this.transformUserData = ({
 			id,
 			hashed_password,
@@ -101,13 +136,13 @@ export class Auth<C extends Configurations = any> {
 					userId: id
 				} as const;
 			};
-			const transform = configs.transformUserData ?? defaultTransform;
+			const transform = config.transformUserData ?? defaultTransform;
 			return transform({ id, ...attributes }) as User;
 		};
-		this.sessionCookie = configs.sessionCookie ?? [defaultSessionCookieOption];
+		this.sessionCookie = config.sessionCookie ?? [defaultSessionCookieOption];
 		this.hash = {
-			generate: configs.hash?.generate ?? generateHashWithScrypt,
-			validate: configs.hash?.validate ?? validateScryptHash
+			generate: config.hash?.generate ?? generateHashWithScrypt,
+			validate: config.hash?.validate ?? validateScryptHash
 		};
 	}
 	public getUser = async (userId: string): Promise<User> => {
@@ -116,6 +151,7 @@ export class Auth<C extends Configurations = any> {
 		const user = this.transformUserData(databaseUser);
 		return user;
 	};
+
 	public getSessionUser = async (
 		sessionId: string
 	): Promise<{
@@ -139,7 +175,7 @@ export class Auth<C extends Configurations = any> {
 				: null;
 		}
 		if (!sessionData) throw new LuciaError("AUTH_INVALID_SESSION_ID");
-		const session = transformDatabaseSessionData(sessionData);
+		const session = validateDatabaseSessionData(sessionData);
 		if (!session) {
 			if (this.autoDatabaseCleanup) {
 				await this.adapter.deleteSession(sessionId);
@@ -152,8 +188,9 @@ export class Auth<C extends Configurations = any> {
 			session
 		};
 	};
+
 	public createUser = async (data: {
-		key: {
+		primaryKey: {
 			providerId: string;
 			providerUserId: string;
 			password: string | null;
@@ -162,13 +199,13 @@ export class Auth<C extends Configurations = any> {
 	}): Promise<User> => {
 		const userId = await this.generateUserId();
 		const userAttributes = data.attributes ?? {};
-		if (data.key === null) {
+		if (data.primaryKey === null) {
 			const userData = await this.adapter.setUser(userId, userAttributes, null);
 			const user = this.transformUserData(userData);
 			return user;
 		}
-		const keyId = `${data.key.providerId}:${data.key.providerUserId}`;
-		const password = data.key.password;
+		const keyId = `${data.primaryKey.providerId}:${data.primaryKey.providerUserId}`;
+		const password = data.primaryKey.password;
 		const hashedPassword = password ? await this.hash.generate(password) : null;
 		const userData = await this.adapter.setUser(userId, userAttributes, {
 			id: keyId,
@@ -178,16 +215,9 @@ export class Auth<C extends Configurations = any> {
 			expires: null
 		});
 		const user = this.transformUserData(userData);
-		const key = transformDatabaseKeyData({
-			id: keyId,
-			user_id: userId,
-			hashed_password: hashedPassword,
-			primary: true,
-			expires: null
-		});
-		if (!key) throw new LuciaError("AUTH_INVALID_KEY_ID");
 		return user;
 	};
+
 	public updateUserAttributes = async (
 		userId: string,
 		attributes: Partial<Lucia.UserAttributes>
@@ -201,35 +231,63 @@ export class Auth<C extends Configurations = any> {
 		const user = this.transformUserData(userData);
 		return user;
 	};
+
 	public deleteUser = async (userId: string): Promise<void> => {
 		await this.adapter.deleteSessionsByUserId(userId);
 		await this.adapter.deleteKeysByUserId(userId);
 		await this.adapter.deleteUser(userId);
 	};
-	public validateKeyPassword = async (
+
+	public useKey = async (
 		providerId: string,
 		providerUserId: string,
-		password: string
+		password: string | null
 	): Promise<Key> => {
 		const keyId = `${providerId}:${providerUserId}`;
-		const databaseKeyData = await this.adapter.getKey(keyId);
+
+		const shouldDataBeDeleted = async (data: KeySchema) => {
+			const isPersistent = data.expires === null;
+			if (isPersistent) return false;
+
+			if (data.hashed_password === null) return true;
+			if (password === null) return false;
+
+			return await this.hash.validate(password, data.hashed_password);
+		};
+
+		const databaseKeyData = await this.adapter.getKey(
+			keyId,
+			shouldDataBeDeleted
+		);
 		if (!databaseKeyData) throw new LuciaError("AUTH_INVALID_KEY_ID");
+		const isSingleUse = databaseKeyData.expires;
+		if (isSingleUse) {
+			const isValid = isWithinExpiration(databaseKeyData.expires);
+			if (!isValid) throw new LuciaError("AUTH_EXPIRED_KEY");
+		}
 		const hashedPassword = databaseKeyData.hashed_password;
-		if (!hashedPassword) throw new LuciaError("AUTH_INVALID_PASSWORD");
-		if (hashedPassword.startsWith("$2a"))
-			throw new LuciaError("AUTH_OUTDATED_PASSWORD");
-		const isValidPassword = await this.hash.validate(password, hashedPassword);
-		if (!isValidPassword) throw new LuciaError("AUTH_INVALID_PASSWORD");
-		const key = transformDatabaseKeyData(databaseKeyData);
-		if (!key) throw new LuciaError("AUTH_INVALID_KEY_ID");
+		if (hashedPassword) {
+			if (!password) throw new LuciaError("AUTH_INVALID_PASSWORD");
+			if (!hashedPassword) throw new LuciaError("AUTH_INVALID_PASSWORD");
+			if (hashedPassword.startsWith("$2a"))
+				throw new LuciaError("AUTH_OUTDATED_PASSWORD");
+			const isValidPassword = await this.hash.validate(
+				password,
+				hashedPassword
+			);
+			if (!isValidPassword) throw new LuciaError("AUTH_INVALID_PASSWORD");
+		}
+
+		const key = transformDatabaseKey(databaseKeyData);
 		return key;
 	};
+
 	public getSession = async (sessionId: string): Promise<Session> => {
 		if (sessionId.length !== 40)
 			throw new LuciaError("AUTH_INVALID_SESSION_ID");
 		const databaseSession = await this.adapter.getSession(sessionId);
 		if (!databaseSession) throw new LuciaError("AUTH_INVALID_SESSION_ID");
-		const session = transformDatabaseSessionData(databaseSession);
+		const session = validateDatabaseSessionData(databaseSession);
 		if (!session) {
 			if (this.autoDatabaseCleanup) {
 				await this.adapter.deleteSession(sessionId);
@@ -238,11 +296,12 @@ export class Auth<C extends Configurations = any> {
 		}
 		return session;
 	};
+
 	public getAllUserSessions = async (userId: string): Promise<Session[]> => {
 		await this.getUser(userId);
 		const databaseData = await this.adapter.getSessionsByUserId(userId);
 		const storedUserSessions = databaseData.map((val) => {
-			const session = transformDatabaseSessionData(val);
+			const session = validateDatabaseSessionData(val);
 			if (session) {
 				return {
 					isValid: true,
@@ -275,12 +334,14 @@ export class Auth<C extends Configurations = any> {
 			.filter((val): val is Session => val !== null);
 		return validUserSessions;
 	};
+
 	public validateSession = async (sessionId: string): Promise<Session> => {
 		const session = await this.getSession(sessionId);
 		if (session.state === "active") return session;
 		const renewedSession = await this.renewSession(sessionId);
 		return renewedSession;
 	};
+
 	public validateSessionUser = async (
 		sessionId: string
 	): Promise<{ session: Session; user: User }> => {
@@ -292,6 +353,7 @@ export class Auth<C extends Configurations = any> {
 			user
 		};
 	};
+
 	public generateSessionId = (): [
 		sessionId: string,
 		activePeriodExpires: Date,
@@ -306,6 +368,7 @@ export class Auth<C extends Configurations = any> {
 		);
 		return [sessionId, activePeriodExpires, idlePeriodExpires];
 	};
+
 	public createSession = async (userId: string): Promise<Session> => {
 		const [sessionId, activePeriodExpires, idlePeriodExpires] =
 			this.generateSessionId();
@@ -329,12 +392,13 @@ export class Auth<C extends Configurations = any> {
 			isFresh: true
 		};
 	};
+
 	public renewSession = async (sessionId: string): Promise<Session> => {
 		if (sessionId.length !== 40)
 			throw new LuciaError("AUTH_INVALID_SESSION_ID");
 		const sessionData = await this.adapter.getSession(sessionId);
 		if (!sessionData) throw new LuciaError("AUTH_INVALID_SESSION_ID");
-		const session = transformDatabaseSessionData(sessionData);
+		const session = validateDatabaseSessionData(sessionData);
 		if (!session) {
 			if (this.autoDatabaseCleanup) {
 				await this.adapter.deleteSession(sessionId);
@@ -349,12 +413,15 @@ export class Auth<C extends Configurations = any> {
 		]);
 		return renewedSession;
 	};
+
 	public invalidateSession = async (sessionId: string): Promise<void> => {
 		await this.adapter.deleteSession(sessionId);
 	};
+
 	public invalidateAllUserSessions = async (userId: string): Promise<void> => {
 		await this.adapter.deleteSessionsByUserId(userId);
 	};
+
 	public deleteDeadUserSessions = async (userId: string): Promise<void> => {
 		const sessions = await this.adapter.getSessionsByUserId(userId);
 		const currentTime = new Date().getTime();
@@ -364,6 +431,7 @@ export class Auth<C extends Configurations = any> {
 		if (deadSessionIds.length === 0) return;
 		await this.adapter.deleteSession(...deadSessionIds);
 	};
+
 	public validateRequestHeaders = (request: MinimalRequest): string => {
 		const cookies = parseCookie(request.headers.get("cookie") ?? "");
 		const sessionId = cookies.auth_session ?? "";
@@ -376,90 +444,115 @@ export class Auth<C extends Configurations = any> {
 		}
 		return sessionId;
 	};
+
 	public createSessionCookies = (session: Session | null): Cookie[] => {
 		return this.sessionCookie.map((options) =>
 			createSessionCookie(session, this.ENV, options)
 		);
 	};
-	public createKey = async (
+
+	public createKey = async <
+		KeyData extends
+			| {
+					readonly type: PersistentKey["type"];
+					providerId: string;
+					providerUserId: string;
+					password: string | null;
+			  }
+			| {
+					readonly type: SingleUseKey["type"];
+					providerId: string;
+					providerUserId: string;
+					password: string | null;
+					timeout: number;
+			  }
+	>(
 		userId: string,
-		keyData: {
-			providerId: string;
-			providerUserId: string;
-			password: string | null;
-			timeout?: number | null;
-		}
-	): Promise<Key> => {
+		keyData: KeyData
+	): Promise<
+		KeyData["type"] extends PersistentKey["type"] ? PersistentKey : SingleUseKey
+	> => {
 		const keyId = `${keyData.providerId}:${keyData.providerUserId}`;
 		const hashedPassword = keyData.password
 			? await this.hash.generate(keyData.password)
 			: null;
+		if (keyData.type === "persistent") {
+			await this.adapter.setKey({
+				id: keyId,
+				user_id: userId,
+				hashed_password: hashedPassword,
+				primary: false,
+				expires: null
+			});
+			return {
+				type: "persistent",
+				providerId: keyData.providerId,
+				providerUserId: keyData.providerUserId,
+				isPrimary: false,
+				isPasswordDefined: !!keyData.password,
+				userId
+			} satisfies PersistentKey as any;
+		}
 		const oneTimeExpires = getOneTimeKeyExpiration(keyData.timeout);
+		if (!oneTimeExpires) throw new TypeError();
 		await this.adapter.setKey({
 			id: keyId,
 			user_id: userId,
 			hashed_password: hashedPassword,
 			primary: false,
-			expires: oneTimeExpires?.getTime() ?? null
+			expires: oneTimeExpires.getTime()
 		});
 		return {
+			type: "single_use",
 			providerId: keyData.providerId,
 			providerUserId: keyData.providerUserId,
-			isPrimary: false,
-			isPasswordDefined: !!keyData.password,
 			userId,
-			oneTimeExpires: oneTimeExpires ?? null
-		};
+			expires: oneTimeExpires,
+			isExpired: !isWithinExpiration(keyData.timeout),
+			isPasswordDefined: !!keyData.password
+		} satisfies SingleUseKey as any;
 	};
+
 	public deleteKey = async (providerId: string, providerUserId: string) => {
 		const keyId = `${providerId}:${providerUserId}`;
 		await this.adapter.deleteNonPrimaryKey(keyId);
 	};
+
 	public getKey = async (
 		providerId: string,
 		providerUserId: string
 	): Promise<Key> => {
 		const keyId = `${providerId}:${providerUserId}`;
-		const keyData = await this.adapter.getKey(keyId);
+		const shouldDataBeDeleted = async () => false;
+		const keyData = await this.adapter.getKey(keyId, shouldDataBeDeleted);
 		if (!keyData) throw new LuciaError("AUTH_INVALID_KEY_ID");
-		const key = transformDatabaseKeyData(keyData);
-		if (!key) throw new LuciaError("AUTH_INVALID_KEY_ID");
+		const key = transformDatabaseKey(keyData);
 		return key;
 	};
 
-	public getKeyUser = async (
-		providerId: string,
-		providerUserId: string
-	): Promise<{
-		key: Key;
-		user: User;
-	}> => {
-		const key = await this.getKey(providerId, providerUserId);
-		const user = await this.getUser(key.userId);
-		return {
-			key,
-			user
-		};
-	};
-	public getAllUserKeys = async (userId: string) => {
+	public getAllUserKeys = async (userId: string): Promise<Key[]> => {
 		await this.getUser(userId);
 		const databaseData = await this.adapter.getKeysByUserId(userId);
-		return databaseData.map((val) => transformDatabaseKeyData(val));
+		return databaseData.map((val) => transformDatabaseKey(val));
 	};
+
 	public updateKeyPassword = async (
 		providerId: string,
 		providerUserId: string,
 		password: string | null
 	): Promise<void> => {
 		const keyId = `${providerId}:${providerUserId}`;
-		const hashedPassword = password ? await this.hash.generate(password) : null;
+		if (password === null) {
+			return await this.adapter.updateKeyPassword(keyId, null);
+		}
+		const hashedPassword = await this.hash.generate(password);
 		await this.adapter.updateKeyPassword(keyId, hashedPassword);
 	};
 }
 
 type MaybePromise<T> = T | Promise<T>;
 
-export interface Configurations {
+export type Configuration = {
 	adapter:
 		| ((E: LuciaErrorConstructor) => Adapter)
 		| {
@@ -480,4 +573,4 @@ export interface Configurations {
 		validate: (s: string, hash: string) => MaybePromise<boolean>;
 	};
 	autoDatabaseCleanup?: boolean;
-}
+};