+++ packages/integration-sveltekit/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @lucia-auth/sveltekit
 
+## 0.6.11
+
+### Patch changes
+
+- [#424](https://github.com/pilcrowOnPaper/lucia/pull/424) by [@pilcrowOnPaper](https://github.com/pilcrowOnPaper) : - Update dependencies
+
+## 0.6.10
+
+### Patch changes
+
+- [#398](https://github.com/pilcrowOnPaper/lucia/pull/398) by [@pilcrowOnPaper](https://github.com/pilcrowOnPaper) : Update peer dependency
+
+## 0.6.9
+
+### Patch changes
+
+- [#395](https://github.com/pilcrowOnPaper/lucia/pull/395) by [@pilcrowOnPaper](https://github.com/pilcrowOnPaper) : Fix types for `handleHooks()`
+
+## 0.6.8
+
+### Patch changes
+
+- [#392](https://github.com/pilcrowOnPaper/lucia/pull/392) by [@pilcrowOnPaper](https://github.com/pilcrowOnPaper) : Update peer dependency
+
+## 0.6.7
+
+### Patch changes
+
+- [#388](https://github.com/pilcrowOnPaper/lucia/pull/388) by [@pilcrowOnPaper](https://github.com/pilcrowOnPaper) : remove unnecessary code
+
 ## 0.6.6
 
 ### Patch changes