+++ packages/lucia-auth/CHANGELOG.md
@@ -1,5 +1,53 @@
 # lucia-auth
 
+## 0.10.0
+
+### Minor changes
+
+- [#424](https://github.com/pilcrowOnPaper/lucia/pull/424) by [@pilcrowOnPaper](https://github.com/pilcrowOnPaper) : Fix issues with single use keys
+
+  - [Breaking] Update `createKey()`
+
+- [#424](https://github.com/pilcrowOnPaper/lucia/pull/424) by [@pilcrowOnPaper](https://github.com/pilcrowOnPaper) : Replace `Token.isExpired()` with `Token.isExpired`
+
+## 0.9.1
+
+### Patch changes
+
+- [#415](https://github.com/pilcrowOnPaper/lucia/pull/415) by [@pilcrowOnPaper](https://github.com/pilcrowOnPaper) : Fix adapter config
+
+## 0.9.0
+
+### Minor changes
+
+- [#398](https://github.com/pilcrowOnPaper/lucia/pull/398) by [@pilcrowOnPaper](https://github.com/pilcrowOnPaper) : **Breaking changes** Better integrate single use keys and clean up APIÏ
+
+  - Update `Key` type
+
+  - Remove `getKeyUser()`
+
+  - Replace `validateKeyPassword()` with `useKey()`
+
+  - Replace `data.key` with `data.primaryKey` for `createUser()`
+
+  - Update `createKey()`
+
+  - `getAllUserKeys()` returns all keys, including those expired
+
+## 0.8.0
+
+### Minor changes
+
+- [#392](https://github.com/pilcrowOnPaper/lucia/pull/392) by [@pilcrowOnPaper](https://github.com/pilcrowOnPaper) : [Breaking] Rename type `Configurations` to `Configuration`
+
+  [Breaking] Attempting to use an expired key throws `AUTH_EXPIRED_KEY`
+
+## 0.7.3
+
+### Patch changes
+
+- [#388](https://github.com/pilcrowOnPaper/lucia/pull/388) by [@pilcrowOnPaper](https://github.com/pilcrowOnPaper) : remove unnecessary code
+
 ## 0.7.2
 
 ### Patch changes